rename binary_search_trees directory to just trees. Add solution for find the winning player in coin game.
