+++ src/math/easy/find_the_winning_player_in_coin_game/mod.rs
@@ -1,0 +1,4 @@
+pub mod mathy;
+
+#[cfg(test)]
+mod tests;
