+++ src/math/easy/mod.rs
@@ -1,1 +1,2 @@
+pub mod find_the_winning_player_in_coin_game;
 pub mod nim_game;
