+++ src/trees/easy/closest_binary_search_tree_value/greedy_search.rs
